+++ Example/myWeb3Wallet/myWeb3Wallet/ViewControllers/WalletViewController/WalletViewController.swift
@@ -11,7 +11,7 @@ import BigInt
 
 final class WalletViewController: UIViewController {
     private let walletManager: WalletManager
-    
+
     private let sendButton: UIButton = {
         let button = UIButton()
         button.setTitle("Send", for: .normal)
@@ -20,7 +20,7 @@ final class WalletViewController: UIViewController {
         button.layer.cornerRadius = 10
         return button
     }()
-    
+
     private let tableView: UITableView = {
         let tableView = UITableView()
         tableView.register(TokenCell.self, forCellReuseIdentifier: "TokenCell")
@@ -31,43 +31,43 @@ final class WalletViewController: UIViewController {
         tableView.refreshControl = UIRefreshControl()
         return tableView
     }()
-    
+
     init(walletManager: WalletManager) {
         self.walletManager = walletManager
         super.init(nibName: nil, bundle: nil)
     }
-    
+
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        
+
         [tableView, sendButton].forEach {
             view.addSubview($0)
             $0.translatesAutoresizingMaskIntoConstraints = false
         }
-        
+
         tableView.bottomAnchor.constraint(equalTo: sendButton.topAnchor).isActive = true
         tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
         tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
         tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
-        
+
         sendButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor,
                                            constant: -16).isActive = true
         sendButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
         sendButton.widthAnchor.constraint(equalTo: view.widthAnchor, multiplier: 0.3).isActive = true
-        
+
         tableView.delegate = self
         tableView.dataSource = self
         tableView.refreshControl?.addTarget(self, action: #selector(loadBalances), for: .valueChanged)
         loadBalances()
-        
+
         sendButton.addTarget(self, action: #selector(send), for: .touchUpInside)
     }
-    
+
     @objc func send() {
         selectChain()
     }
@@ -79,13 +79,13 @@ final class WalletViewController: UIViewController {
             }
             alertController.addAction(action)
         }
-        
+
         let action = UIAlertAction.init(title: "Cancel", style: .destructive)
         alertController.addAction(action)
-        
+
         self.present(alertController, animated: true)
     }
-    
+
     func selectToken(network: Web3Network) {
         let alertController = UIAlertController(title: "Select token", message: "", preferredStyle: .actionSheet)
         for token in network.network.tokens {
@@ -94,13 +94,13 @@ final class WalletViewController: UIViewController {
             }
             alertController.addAction(action)
         }
-        
+
         let action = UIAlertAction.init(title: "Cancel", style: .destructive)
         alertController.addAction(action)
-        
+
         self.present(alertController, animated: true)
     }
-    
+
     func sendToken(_ token: Token, from network: Web3Network) {
         let sendVC = SendViewController(walletManager: walletManager, network: network, token: token)
         navigationController?.pushViewController(sendVC, animated: true)
@@ -111,6 +111,7 @@ extension WalletViewController {
     @objc func loadBalances() {
         Task { @MainActor in
             await walletManager.loadBalances()
+            tableView.refreshControl?.endRefreshing()
             tableView.reloadData()
         }
     }
@@ -124,7 +125,7 @@ extension WalletViewController: UITableViewDelegate, UITableViewDataSource {
         if section == 0 { return 1 }
         return walletManager.networks[section - 1].network.tokens.count
     }
-    
+
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         if indexPath.section == 0 {
             guard let cell = tableView.dequeueReusableCell(withIdentifier: "WalletCell", for: indexPath) as? WalletCell
@@ -138,7 +139,7 @@ extension WalletViewController: UITableViewDelegate, UITableViewDataSource {
             let token = network.network.tokens[indexPath.row]
             cell.network = network.network.name
             cell.name = token.symbol
-            
+
             if let balance = network.tokensBalances[token.symbol] {
                 let doubleString = Utilities.formatToPrecision(balance, units: .custom(token.decimals),
                                                                formattingDecimals: 10)
@@ -148,6 +149,6 @@ extension WalletViewController: UITableViewDelegate, UITableViewDataSource {
             }
             return cell
         }
-        
+
     }
 }
